+++ src/main/java/com/qph/facturacion/repository/FacturaRepository.java
@@ -3,5 +3,5 @@ package com.qph.facturacion.repository;
 import com.qph.facturacion.entity.Factura;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FacturaRepository extends JpaRepository<Factura,Integer> {
+public interface FacturaRepository extends JpaRepository<Factura,Long> {
 }
